+++ apps/frontend/src/app/5_pages/ZeroPage/ZeroPage.tsx
@@ -57,6 +57,15 @@ export const ZeroPage: FC = () => {
   const { connectWallet } = useWalletConnect();
   const { account } = useAccount();
 
+  const { data: userOpenTrove } = useGetUserOpenTrove();
+
+  const isUserOpenTrove = useMemo(() => {
+    if (account) {
+      return userOpenTrove?.trove?.changes[0]?.trove.status === 'open';
+    }
+    return false;
+  }, [userOpenTrove, account]);
+
   const collateral = useMemo(
     () => Number(trove?.collateral ?? 0),
     [trove?.collateral],
@@ -147,7 +156,10 @@ export const ZeroPage: FC = () => {
                     onWithdraw={claimCollateralSurplus}
                   />
                   <div className="h-80 md:flex-1 bg-gray-80 rounded pt-2 pr-2 flex items-center">
-                    <LOCChart />
+                    <LOCChart
+                      userOpenTrove={userOpenTrove}
+                      isUserOpenTrove={isUserOpenTrove}
+                    />
                   </div>
                 </>
               )}
@@ -205,7 +217,10 @@ export const ZeroPage: FC = () => {
                   </Paragraph>
 
                   <div className="h-80 md:flex-1 bg-gray-80 rounded pt-2 pr-2 flex items-center">
-                    <LOCChart />
+                    <LOCChart
+                      userOpenTrove={userOpenTrove}
+                      isUserOpenTrove={isUserOpenTrove}
+                    />
                   </div>
                 </div>
               </div>
